Raft (wip)

+++ raft/term.go
@@ -9,7 +9,7 @@ import (
 
 // A term represents a particular member state in the Raft epochal time model.
 type term struct {
-	Num      int        `json:"num"`       // the current term number (increases monotonically across the cluster)
+	Num      int64      `json:"num"`       // the current term number (increases monotonically across the cluster)
 	LeaderId *uuid.UUID `json:"leader_id"` // the current leader (as seen by this member)
 	VotedFor *uuid.UUID `json:"voted_for"` // who was voted for this term (guaranteed not nil when leader != nil)
 }
